[not verified] my-jetpack: refactor -> move purchases endpoint to its own file
